Moved main.c to old and abandomed old todo
created old dir and moved main.c to it. Abandomed old cli-todo app because the catagory system nedds to be reworked and the code could also be rewritten
