--- CLI-Todo/add.txt
@@ -1,12 +1,0 @@
-categorys (like Ferien: ) highlit with maybe Backround colors
-
--add category
-	Enter Category (dannach scanf(...)
--rm category 
-
-bei list dann if(anfang ="CATEGORY" dann ignorieren)
-
-rename
-sync function
-
-
